+++ Communication/Packets/Outgoing/Navigator/New/NavigatorSearchResultSetComposer.cs
@@ -1,10 +1,12 @@
 ﻿using Plus.HabboHotel.GameClients;
 using Plus.HabboHotel.Navigator;
+using Plus.HabboHotel.Rooms;
 
 namespace Plus.Communication.Packets.Outgoing.Navigator.New;
 
 public class NavigatorSearchResultSetComposer : IServerPacket
 {
+    private readonly INavigatorQueryService _navigatorQueryService;
     private readonly string _category;
     private readonly string _data;
     private readonly ICollection<SearchResultList> _searchResultLists;
@@ -14,12 +16,13 @@ public class NavigatorSearchResultSetComposer : IServerPacket
     public uint MessageId => ServerPacketHeader.NavigatorSearchResultSetComposer;
 
     public NavigatorSearchResultSetComposer(string category, string data,
-        ICollection<SearchResultList> searchResultLists, GameClient session, int goBack = 1, int fetchLimit = 12)
+        ICollection<SearchResultList> searchResultLists, GameClient session, INavigatorQueryService navigatorQueryService, int goBack = 1, int fetchLimit = 12)
     {
         _category = category;
         _data = data;
         _searchResultLists = searchResultLists;
         _session = session;
+        _navigatorQueryService = navigatorQueryService;
         _goBack = goBack;
         _fetchLimit = fetchLimit;
     }
@@ -41,7 +44,10 @@ public class NavigatorSearchResultSetComposer : IServerPacket
             packet.WriteInteger(searchResult.ViewMode == NavigatorViewMode.Regular ? 0 :
                 searchResult.ViewMode == NavigatorViewMode.Thumbnail ? 1 :
                 0); //View mode, 0 = tiny/regular, 1 = thumbnail
-            NavigatorHandler.Search(packet, searchResult, _data, _session, _fetchLimit);
+            var rooms = _navigatorQueryService.GetSearchResults(searchResult, _data, _session, _fetchLimit);
+            packet.WriteInteger(rooms.Count);
+            foreach (var room in rooms)
+                RoomAppender.WriteRoom(packet, room, room.Promotion);
         }
     }
 }