+++ HabboHotel/Navigator/INavigatorQueryService.cs
@@ -1,0 +1,9 @@
+using Plus.HabboHotel.GameClients;
+using Plus.HabboHotel.Rooms;
+
+namespace Plus.HabboHotel.Navigator;
+
+public interface INavigatorQueryService
+{
+    ICollection<RoomData> GetSearchResults(SearchResultList result, string query, GameClient session, int limit);
+}